Main
The Library Android App is a robust solution for students to access and manage library resources from their mobile devices. It provides secure login and profile management features, access to a wide range of library materials, and the ability to view PDFs directly within the app. The backend, powered by Express and Node.js, ensures efficient handling of data and authentication, while MongoDB serves as the reliable database. This integration of technologies offers a seamless and user-friendly experience for students, enhancing their interaction with the library's digital resources.